# backend/openapi.yaml
openapi: 3.1.0
info:
  title: "TinyChat API"
  version: 0.1.0
  description: "API contract for the TinyChat backend."
security:
  - bearerAuth: []
paths:
...
  /api/billing/rates:
    get:
      operationId: getBillingRates
      summary: Public per-model credit rates
      description: >-
        Public (unauthenticated) credit-rate table. Denominated in credits only —
        no dollar/USD fields appear in this payload. Served from the shared
        RedPill catalog and mirrors `/api/chat/models`' 500/502 error contract.
      security: []
      responses:
        "200":
          description: Baseline model id and per-model rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatesResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /api/billing/status:
    get:
      operationId: getBillingStatus
      summary: Caller's tier, usage, and subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Billing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStatus"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/billing/checkout:
    post:
      operationId: createCheckout
      summary: Create a Stripe Checkout session for a paid tier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/BillingNotConfigured"
  /api/billing/portal:
    post:
      operationId: createBillingPortal
      summary: Create a Stripe billing portal session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "503":
          $ref: "#/components/responses/BillingNotConfigured"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend session token returned by SIWE verification.
  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...
    RateInfo:
      type: object
      required: [id, creditsPerKInput, creditsPerKOutput, multiplier]
      properties:
        id:
          type: string
        creditsPerKInput:
          type: number
        creditsPerKOutput:
          type: number
        multiplier:
          type: number
          description: Picker-badge multiplier relative to the baseline model's output rate.
    RatesResponse:
      type: object
      required: [baseline, models]
      properties:
        baseline:
          type: string
          description: Model id used as the multiplier baseline (REDPILL_DEFAULT_MODEL).
        models:
          type: array
          items:
            $ref: "#/components/schemas/RateInfo"
...
    UsageInfo:
      type: object
      required: [used, limit, resetsAt]
      properties:
        used:
          type: integer
        limit:
          type: integer
        resetsAt:
          type: string
          format: date-time
          description: >-
            Next budget-window boundary. Free: the next UTC day boundary. Paid:
            the next anchored-week boundary (subscriber's billing-cycle anchor
            plus an integer number of 7-day periods).
    SubscriptionInfo:
      type: object
      required: [status, interval, currentPeriodEnd, anchor]
      properties:
        status:
          type: string
        interval:
          type:
            - string
            - "null"
          enum: [monthly, yearly, null]
        currentPeriodEnd:
          type:
            - string
            - "null"
          format: date-time
        anchor:
          type:
            - string
            - "null"
          format: date-time
          description: >-
            Stripe billing-cycle anchor (subscription-stable across renewals).
            Paid tiers' weekly windows are measured as 7-day periods from this
            instant. Null for free tier.
    BillingStatus:
      type: object
      required: [tier, usage, subscription]
      properties:
        tier:
          type: string
          enum: [free, plus, pro]
        usage:
          $ref: "#/components/schemas/UsageInfo"
        subscription:
          oneOf:
            - $ref: "#/components/schemas/SubscriptionInfo"
            - type: "null"
    CheckoutRequest:
      type: object
      required: [tier, interval]
      properties:
        tier:
          type: string
          enum: [plus, pro]
        interval:
          type: string
          enum: [monthly, yearly]
    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri

    # ── Connector webhooks ──────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthenticated:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    UpstreamError:
      description: Upstream LLM API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    BillingNotConfigured:
      description: Billing/Stripe is not configured on this server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...